feat: update following feedbacks
