+++ pkg/processor/doc.go
@@ -12,5 +12,64 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// Package processor provides utilities for managing game server client processes
+// Package processor provides client functionality for the Agones allocation processor system
+//
+// The allocation processor system enables batched game server allocation requests
+// This package contains the client implementation that connects to the processor service
+// batches allocation requests, and handles the stream-based communication protocol
+//
+// Key components:
+// - Client: Manages connection lifecycle and request batching
+// - Config: Configuration for processor client behavior
+// - Batch handling: Accumulates requests and sends them in batches to the processor
+//
+// The client establishes a bidirectional gRPC stream with the processor service,
+// registers itself, and then handles pull requests (to send batched allocations)
+// and batch responses (containing allocation results)
+//
+// Flow diagram:
+//
+//	┌─────────────────┐    ┌─────────────────────┐    ┌─────────────────────┐
+//	│     Client      │    │  Processor Client   │    │  Processor Server   │
+//	│  (Allocator/    │    │   (this package)    │    │                     │
+//	│   Extension)    │    │                     │    │                     │
+//	└─────────────────┘    └─────────────────────┘    └─────────────────────┘
+//	         │                        │                          │
+//	         │                        │ 1. Connect & Register    │
+//	         │                        │    (bidirectional stream)│
+//	         │                        ├──────────────────────────►
+//	         │                        │                          │
+//	         │ 2. Allocate(request)   │                          │
+//	         ├────────────────────────►                          │
+//	         │                        │                          │
+//	         │                        │ 3. Add to hotBatch       │
+//	         │                        │    (accumulate)          │
+//	         │                        │                          │
+//	         │                        │ 4. Pull Request          │
+//	         │                        │    (to all clients)      │
+//	         │                        ◄──────────────────────────┤
+//	         │                        │ 5. Send BatchRequest     │
+//	         │                        │    (hotBatch)            │
+//	         │                        ├──────────────────────────►
+//	         │                        │                          │
+//	         │                        │                          │ 6. Process
+//	         │                        │                          │    allocations
+//	         │                        │                          │
+//	         │                        │ 7. BatchResponse         │
+//	         │                        │    (results)             │
+//	         │                        ◄──────────────────────────┤
+//	         │ 8. Return result       │                          │
+//	         ◄────────────────────────┤                          │
+//	         │                        │                          │
+//
+//	Note: Multiple Processor Clients can connect to one Processor Server
+//	      The server sends pull requests to all connected clients
+//
+//	Legend:
+//	- Client: Makes allocation requests (allocator, extensions, etc.)
+//	- Processor Client: Batches requests and manages communication
+//	- Processor Server: Processes batched allocations from multiple clients
+//	- Pull Request: Server asks all connected clients for pending requests
+//	- BatchRequest: Client sends accumulated allocation requests
+//	- BatchResponse: Server returns allocation results
 package processor
